🚀 feat(main.rs): Implement password checking functionality using the pwnedpasswords API
A new `PasswordStatus` enum and `PasswordCheckResult` struct were added to represent the status of a password's security. The `PasswordChecker` struct was also created to house the `check_password` function.

The `check_password` function sends a request to the pwnedpasswords API with the first five characters of the SHA-1 hash of the password to check if it has been compromised. If it has been compromised, the function returns the count of how many times it has been found in previous data breaches, otherwise it returns that the password is safe.

The `PasswordCheckResult` struct holds the password and its `PasswordStatus`. The `PasswordStatus` enum has two variants, `Safe` and `Compromised(u32)` which holds the count of how many times the password has been found in previous data breaches.

Changes made:
- Added `PasswordStatus` enum
- Added `PasswordCheckResult` struct
- Added `PasswordChecker` struct with `check_password` function
- Modified the `main` function to use `check_password` function and print results
Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>